# openapi.yaml
openapi: 3.1.0
info:
  title: UARE Custom Backend API
  version: 1.0.0
  description: |
    Production API for the UARE enterprise platform — auth, billing, orgs,
    CAD execution, workflows, workers, analytics, patents, and ops quality.
servers:
  - url: http://localhost:8787
    description: Local development
  - url: https://api.uare.app
    description: Production
security:
  - cookieAuth: []
  - devHeaderAuth: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: uare_session
    devHeaderAuth:
      type: apiKey
      in: header
      name: x-dev-user-id
    opsBearerToken:
      type: http
      scheme: bearer
  schemas:
...
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          example: true
...
    Org:
      type: object
      properties:
        org_id:
          type: string
        slug:
          type: string
        name:
          type: string
        owner_user_id:
          type: string
        created_at:
          type: string
          format: date-time
    Membership:
      type: object
      properties:
        membership_id:
          type: string
        org_id:
          type: string
        user_id:
          type: string
        role:
          type: string
          enum: [owner, admin, member, billing_manager, viewer]
        created_at:
          type: string
          format: date-time
    AuditEvent:
      type: object
      properties:
        audit_id:
          type: string
        actor_id:
          type: string
          nullable: true
        actor_role:
          type: string
          nullable: true
        action:
          type: string
        target_type:
          type: string
          nullable: true
        target_id:
          type: string
          nullable: true
        status:
          type: string
        request_id:
          type: string
          nullable: true
        ip:
          type: string
          nullable: true
        metadata_json:
          type: object
        created_at:
          type: string
          format: date-time
...
paths:
...
  /orgs:
    get:
      tags: [orgs]
      summary: List orgs for the current user
      responses:
        '200':
          description: Orgs list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  orgs:
                    type: array
                    items:
                      $ref: '#/components/schemas/Org'
    post:
      tags: [orgs]
      summary: Create a new organization
      parameters:
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 200
                slug:
                  type: string
                  maxLength: 100
      responses:
        '201':
          description: Org created
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  org:
                    $ref: '#/components/schemas/Org'
  /orgs/{orgId}/members:
    get:
      tags: [orgs]
      summary: List members of an organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Members list
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  members:
                    type: array
                    items:
                      $ref: '#/components/schemas/Membership'
    post:
      tags: [orgs]
      summary: Add or update a member in an organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: header
          name: Idempotency-Key
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                user_id:
                  type: string
                role:
                  type: string
                  enum: [owner, admin, member, billing_manager, viewer]
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  membership:
                    $ref: '#/components/schemas/Membership'
  /orgs/{orgId}/members/{userId}:
    delete:
      tags: [orgs]
      summary: Remove a member from an organization
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: path
          name: userId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Member removed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
  /orgs/{orgId}/audit:
    get:
      tags: [orgs]
      summary: View org-scoped audit log (owner/admin only)
      parameters:
        - in: path
          name: orgId
          required: true
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            maximum: 200
            default: 50
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: actor_id
          schema:
            type: string
        - in: query
          name: target_type
          schema:
            type: string
        - in: query
          name: target_id
          schema:
            type: string
      responses:
        '200':
          description: Audit events scoped to this org
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  org_id:
                    type: string
                  events:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditEvent'
        '403':
          description: Insufficient org role

  # ── Projects ──────────────────────────────────────────────────────────────
...